fix DTO and mappers

+++ data/products/src/main/java/ru/ikom/products/Mappers.kt
@@ -14,7 +14,7 @@ internal fun ProductDTO.toProductData(): ProductData =
         discountPercentage,
         id,
         images,
-        price,
+        price.toInt(),
         rating,
         stock,
         thumbnail,
